Datasaver + read binary logic

+++ ogb/graphproppred/dataset_dgl.py
@@ -6,40 +6,56 @@ import numpy as np
 from dgl.data.utils import load_graphs, save_graphs, Subset
 import dgl
 from ogb.utils.url import decide_download, download_url, extract_zip
-from ogb.io.read_graph_dgl import read_csv_graph_dgl
+from ogb.io.read_graph_dgl import read_graph_dgl
 
 class DglGraphPropPredDataset(object):
     '''Adapted from https://docs.dgl.ai/en/latest/_modules/dgl/data/chem/csv_dataset.html#CSVDataset'''
-    def __init__(self, name, root = 'dataset'):
+    def __init__(self, name, root = 'dataset', meta_dict = None):
+        '''
+            - name (str): name of the dataset
+            - root (str): root directory to store the dataset folder
+
+            - meta_dict: dictionary that stores all the meta-information about data. Default is None, 
+                    but when something is passed, it uses its information. Useful for debugging for external contributers.
+        '''
+
         self.name = name ## original name, e.g., ogbg-molhiv
         
-        self.dir_name = '_'.join(name.split('-')) + '_dgl' ## replace hyphen with underline, e.g., ogbg_molhiv_dgl
-        self.original_root = root
-        self.root = osp.join(root, self.dir_name)
-
-        self.meta_info = pd.read_csv(os.path.join(os.path.dirname(__file__), 'master.csv'), index_col = 0)
-        if not self.name in self.meta_info:
-            print(self.name)
-            error_mssg = 'Invalid dataset name {}.\n'.format(self.name)
-            error_mssg += 'Available datasets are as follows:\n'
-            error_mssg += '\n'.join(self.meta_info.keys())
-            raise ValueError(error_mssg)
+        if meta_dict is None:
+            self.dir_name = '_'.join(name.split('-')) + '_dgl' ## replace hyphen with underline, e.g., ogbg_hiv_dgl
+            self.original_root = root
+            self.root = osp.join(root, self.dir_name)
+            
+            master = pd.read_csv(os.path.join(os.path.dirname(__file__), 'master.csv'), index_col = 0)
+            if not self.name in master:
+                error_mssg = 'Invalid dataset name {}.\n'.format(self.name)
+                error_mssg += 'Available datasets are as follows:\n'
+                error_mssg += '\n'.join(master.keys())
+                raise ValueError(error_mssg)
+            self.meta_info = master[self.name]
+            
+        else:
+            self.dir_name = meta_dict['dir_path']
+            self.original_root = ''
+            self.root = meta_dict['dir_path']
+            self.meta_info = meta_dict
 
         # check version
         # First check whether the dataset has been already downloaded or not.
         # If so, check whether the dataset version is the newest or not.
         # If the dataset is not the newest version, notify this to the user. 
-        if osp.isdir(self.root) and (not osp.exists(osp.join(self.root, 'RELEASE_v' + str(self.meta_info[self.name]['version']) + '.txt'))):
+        if osp.isdir(self.root) and (not osp.exists(osp.join(self.root, 'RELEASE_v' + str(self.meta_info['version']) + '.txt'))):
             print(self.name + ' has been updated.')
             if input('Will you update the dataset now? (y/N)\n').lower() == 'y':
                 shutil.rmtree(self.root)
 
-        self.download_name = self.meta_info[self.name]['download_name'] ## name of downloaded file, e.g., tox21
+        self.download_name = self.meta_info['download_name'] ## name of downloaded file, e.g., tox21
 
-        self.num_tasks = int(self.meta_info[self.name]['num tasks'])
-        self.eval_metric = self.meta_info[self.name]['eval metric']
-        self.task_type = self.meta_info[self.name]['task type']
-        self.num_classes = self.meta_info[self.name]['num classes']
+        self.num_tasks = int(self.meta_info['num tasks'])
+        self.eval_metric = self.meta_info['eval metric']
+        self.task_type = self.meta_info['task type']
+        self.num_classes = self.meta_info['num classes']
+        self.binary = self.meta_info['binary'] == 'True'
 
         super(DglGraphPropPredDataset, self).__init__()
 
@@ -64,37 +80,45 @@ class DglGraphPropPredDataset(object):
                 self.labels = label_dict['labels']
 
         else:
-            ### download
-            url = self.meta_info[self.name]['url']
-            if decide_download(url):
-                path = download_url(url, self.original_root)
-                extract_zip(path, self.original_root)
-                os.unlink(path)
-                # delete folder if there exists
-                try:
-                    shutil.rmtree(self.root)
-                except:
-                    pass
-                shutil.move(osp.join(self.original_root, self.download_name), self.root)
+            ### check download
+            if self.binary:
+                # npz format
+                has_necessary_file = osp.exists(osp.join(self.root, 'raw', 'data.npz'))
             else:
-                print('Stop download.')
-                exit(-1)
+                # csv file
+                has_necessary_file = osp.exists(osp.join(self.root, 'raw', 'edge.csv.gz'))
+            
+            ### download
+            if not has_necessary_file:
+                url = self.meta_info['url']
+                if decide_download(url):
+                    path = download_url(url, self.original_root)
+                    extract_zip(path, self.original_root)
+                    os.unlink(path)
+                    # delete folder if there exists
+                    try:
+                        shutil.rmtree(self.root)
+                    except:
+                        pass
+                    shutil.move(osp.join(self.original_root, self.download_name), self.root)
+                else:
+                    print('Stop download.')
+                    exit(-1)
 
             ### preprocess
-            add_inverse_edge = self.meta_info[self.name]['add_inverse_edge'] == 'True'
+            add_inverse_edge = self.meta_info['add_inverse_edge'] == 'True'
 
-            if self.meta_info[self.name]['additional node files'] == 'None':
+            if self.meta_info['additional node files'] == 'None':
                 additional_node_files = []
             else:
-                additional_node_files = self.meta_info[self.name]['additional node files'].split(',')
+                additional_node_files = self.meta_info['additional node files'].split(',')
 
-            if self.meta_info[self.name]['additional edge files'] == 'None':
+            if self.meta_info['additional edge files'] == 'None':
                 additional_edge_files = []
             else:
-                additional_edge_files = self.meta_info[self.name]['additional edge files'].split(',')
+                additional_edge_files = self.meta_info['additional edge files'].split(',')
 
-            graphs = read_csv_graph_dgl(raw_dir, add_inverse_edge = add_inverse_edge, additional_node_files = additional_node_files, additional_edge_files = additional_edge_files)
-
+            graphs = read_graph_dgl(raw_dir, add_inverse_edge = add_inverse_edge, additional_node_files = additional_node_files, additional_edge_files = additional_edge_files, binary=self.binary)
 
             if self.task_type == 'subtoken prediction':
                 # the downloaded labels are initially joined by ' '
@@ -111,7 +135,10 @@ class DglGraphPropPredDataset(object):
                 self.labels = torch.load(target_sequence_file_path)
 
             else:
-                labels = pd.read_csv(osp.join(raw_dir, 'graph-label.csv.gz'), compression='gzip', header = None).values
+                if self.binary:
+                    labels = np.load(osp.join(raw_dir, 'graph-label.npz'))['graph_label']
+                else:
+                    labels = pd.read_csv(osp.join(raw_dir, 'graph-label.csv.gz'), compression='gzip', header = None).values
 
                 has_nan = np.isnan(labels).any()
 
@@ -134,7 +161,7 @@ class DglGraphPropPredDataset(object):
 
     def get_idx_split(self, split_type = None):
         if split_type is None:
-            split_type = self.meta_info[self.name]['split']
+            split_type = self.meta_info['split']
             
         path = osp.join(self.root, 'split', split_type)
